mcp/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "product-management"
version = "1.0.0"
description = "A Python library for managing product information with SQLite backend and MCP server support"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Zensar AI for Leaders"}
]
keywords = ["products", "inventory", "management", "mcp", "sqlalchemy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.23",
    "fastmcp>=2.13.1",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.30.0",
    "fastapi>=0.115.0",
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[project.urls]
Homepage = "https://github.com/iomegak12/zensar-trg-sources"
Repository = "https://github.com/iomegak12/zensar-trg-sources"

[project.scripts]
products-server = "run_mcp_server:main"

[tool.setuptools]
packages = ["product_management"]
